Extracted utility functions from data-builder tool
